# Fix pygame.draw.arc