Fix a rendering bug with pygame.draw.arc

This bug already has a pull request open (#1981), however there has been no activity on that PR for four months (except for a comment asking about its revival a few days ago).

I used the following sources for testing code: https://github.com/pygame/pygame/issues/3020#issuecomment-1064505060 as well as https://github.com/pygame-community/pygame-ce/pull/1981#pullrequestreview-1336063364

I intend to make a few more performance based commits, as there are a number of different ways I can think of to significantly improve performance.
I figured I should open the PR now as I have a working solution and I figured it would be good to do so at this point.

EDIT:

I have finished making performance improvements and am now just waiting for a review before we can merge this PR. I'd be happy to implement any other changes anyone can think of