@Cyuria 2023-07-25
difference:
old:
![pygame old arc](https://github.com/pygame-community/pygame-ce/assets/55673467/a4aae3d4-f0ea-4573-9af1-c45899876d0b)
![pygame_arc_test2_old](https://github.com/pygame-community/pygame-ce/assets/55673467/218b220d-9895-4321-bb2c-2c4036e9f706)
new:
![pygame new arc 2](https://github.com/pygame-community/pygame-ce/assets/55673467/56c6d15e-4dcc-4dae-ab10-56c2e14e95f8)
![pygame_arc_test2_new](https://github.com/pygame-community/pygame-ce/assets/55673467/526a4b77-261f-4436-a81f-93b2febd470b)