@Cyuria 2023-07-25
new tests:
![pygame_arc_test2_new2](https://github.com/pygame-community/pygame-ce/assets/55673467/62b7b579-4d68-4fb6-b3b4-f1c24e8fd6ae)
![pygame new arc 3](https://github.com/pygame-community/pygame-ce/assets/55673467/045d74ea-7916-42be-8a0e-660d41893f1c)