@Cyuria 2023-07-25
Could I get someone else's opinion on this?

The reason the tests are failing is because of some differences between the new arc algorithm and the old one when the difference between the start and stop angles is greater than 2 Pi radians, or 360 degrees.

At this point I will probably add a clause to my algorithm to satisfy the tests.

However, I think my algorithm's way of handling the edge case could be preferable if documented. Here's a comparison:
Old:
![oldpygamearc](https://github.com/pygame-community/pygame-ce/assets/55673467/87b3c215-bcb5-421a-8fd7-efdfcfa6a423)
New:
![pygamearc](https://github.com/pygame-community/pygame-ce/assets/55673467/f16bf1ba-0c6f-49aa-8cff-20d76c3b4d1d)
The code I used to test this:
```py
import math
import time

import pygame

pygame.init()
font = pygame.font.Font(size=32)

def drawarc(screen):
    t = time.time() % (4 * math.pi)
    screen.fill((0, 0, 0))
    pygame.draw.arc(screen, pygame.Color('white'), (50, 50, 300, 300), 0, t, 50)
    screen.blit(font.render(str(round(t / math.pi, 1)) + ' Pi radians', True, pygame.Color('white')), (0, 0))
    pygame.display.flip()

def runanim():
    screen = pygame.display.set_mode((400, 400))
    while True:
        event = pygame.event.poll()
        drawarc(screen)
        if event.type == pygame.NOEVENT:
            continue
        if event.type == pygame.QUIT:
            break
        if event.type == pygame.KEYDOWN:
            print(time.time() % (4 * math.pi))
            continue
        if event.type == pygame.MOUSEBUTTONDOWN:
            print(event.pos)
            continue


if __name__ == "__main__":
    runanim()
```