@oddbookworm 2023-07-25
> The reason the tests are failing is because of some differences between the new arc algorithm and the old one when the difference between the start and stop angles is greater than 2 Pi radians, or 360 degrees.
>
>At this point I will probably add a clause to my algorithm to satisfy the tests.

Generally, we try to avoid changes that might break someone's project. I also would like feedback from other reviewers about which behavior should be the "correct" behavior. IMO if there is a difference of more than 2pi radians, then a full circle should be drawn because it has important information in some cases. I could potentially be persuaded otherwise though