@Cyuria 2023-07-26
I've been working a bit more and I believe I have run into a place where the old algorithm (as well as one of the tests) are either inconsistent with the docs or just broken, if I fix the start angle at -17 radians and the stop angle at -5.5, (the only currently failing test), the following happens:
![pygame arc oldnew compare](https://github.com/pygame-community/pygame-ce/assets/55673467/723e7eea-9bc8-4daf-b6ac-237b4cb68d7c)
Unless I am making a mistake, the documentation says otherwise:
https://github.com/pygame-community/pygame-ce/blob/6ebeb191f5af2927a73a74b2c3388d2c4f4225ab/docs/reST/ref/draw.rst?plain=1#L270-L278