@Cyuria 2023-07-26
seeing as all tests passed, I started working a bit on performance, using the following benchmark script:

```py
import math
import time

import pygame

def drawarc(surface, t):
    pygame.draw.arc(surface, pygame.Color('white'), (0, 0, 1000, 1000), 0, t, 300)

def runbench():
    screen = pygame.display.set_mode((50, 50))
    surface = pygame.Surface((1000, 1000))
    starttime = time.time_ns()

    for c in range(100):
        for i in range(200):
            drawarc(surface, i * math.pi / 100)

        # give the user some kind of feedback
        screen.fill((256 * c // 100,)*3)
        pygame.display.flip()

    endtime = time.time_ns()

    print(f"starttime:  {starttime / 1000000000}s")
    print(f"endtime:    {endtime / 1000000000}s")
    print(f"difference: {(endtime - starttime) / 1000000000}s")

if __name__ == "__main__":
    runbench()

```

The following are my results for the entire test

| algorithm | time |
|-----------|------|
|    old    |  47s |
|  new (1)  |  69s |
|  new (2)  |  54s |

(1) the first full implementation which passed all tests (commit a9522a69a1f18636d700a047f78e78356277915a)
(2) with bounds checking before running the algorithm (commit a31b616ff9e94c5b88192dd43f09c692d0535553)

It is also possible to visualize the bounds checking by adding the following code to the draw_arc function anywhere after the minima and maxima are calculated:
```c
    draw_line(surf, minx + x_center, miny + y_center,
              minx + x_center, maxy + y_center, color, drawn_area);
    draw_line(surf, minx + x_center, maxy + y_center,
              maxx + x_center, maxy + y_center, color, drawn_area);
    draw_line(surf, maxx + x_center, maxy + y_center,
              maxx + x_center, miny + y_center, color, drawn_area);
    draw_line(surf, maxx + x_center, miny + y_center,
              minx + x_center, miny + y_center, color, drawn_area);
```

doing so gives the following result on my machine with the animated test program (albeit with red for the bounding box lines)

![pygamearc_bounding](https://github.com/pygame-community/pygame-ce/assets/55673467/1b5ffc47-7a84-43c0-acd0-37c538532c85)

There are still two more performance improvements I can think of which I will try to implement soon.
- reducing the amount of per pixel checking by using alternate methods to calculate the easier parts of the arc
- directly setting the pixels instead of updating drawn area and using the bounding box given values to calculate the drawn area