@Cyuria 2023-07-27
The latest performance enhancements are documented in the two above commits. Here is the updated benchmark table:

| algorithm | time |
|-----------|------|
|    old    |  47s |
|  new (1)  |  69s |
|  new (2)  |  54s |
|  new (3)  |  48s |
|  new (4)  |  35s |

(1) the first full implementation which passed all tests (commit https://github.com/pygame-community/pygame-ce/commit/a9522a69a1f18636d700a047f78e78356277915a)
(2) with bounds checking before running the algorithm (commit https://github.com/pygame-community/pygame-ce/commit/a31b616ff9e94c5b88192dd43f09c692d0535553)

(3) Use the calculated bounding box to obtain the drawn_area instead of calculating it for each drawn pixel with `set_and_check_rect` (commit 2e3585308911a5c92ac2d89a95c690214f3a9e10) -- NOTE this commit fails the draw arc bounding box test due to an out by one error on some arcs, will be fixed in the next commit
(4) Bypass `set_at` to avoid checking surface clip rect and check it when calculating the bounding box instead (commit bef7f717c44391b37bff85533125a8d2234e5fe1) -- NOTE this commit fails the draw arc clip rect box due to an out by one error, will be fixed in the next commit