@Starbuck5 2023-07-27
So this takes the old algorithm from mostly unusable (because of the holes) to twice as fast? Amazing!

Cyuria, are you on the pygame community discord?