@Cyuria 2023-07-27
I fixed the test cases, here's the updated performance table:

| algorithm | time |
| --------- | ---- |
|    old    |  47s |
|  new (1)  |  69s |
|  new (2)  |  54s |
|  new (3)  |  48s |
|  new (4)  |  35s |
|  new (5)  |  36s |

(5) up to the latest commit with test cases (commit f77a8912e5954c45094bc5ca3efef2400be4db06)

The timings can vary pretty badly, I ran some of the tests a few times and got up to a few seconds difference for each run, so don't necessarily trust the extra +1s (although there definitely should be slightly lower performance because I'm running quite a few extra checks, especially on arcs which are not fully inside the clipping rect of the drawn surface)