@Cyuria 2023-07-29
I think I'm done with this PR unless someone has anything else to say, I still need a review.

The performance table for each commit is as follows:

| algorithm | time |
|-----------|------|
|    old    |  47s |
|  new (1)  |  69s |
|  new (2)  |  54s |
|  new (3)  |  48s |
|  new (5)  |  36s |
|  new (6)  |  30s |

(1) the first full implementation which passed all tests (commit https://github.com/pygame-community/pygame-ce/commit/a9522a69a1f18636d700a047f78e78356277915a)

(2) with bounds checking before running the algorithm (commit https://github.com/pygame-community/pygame-ce/commit/a31b616ff9e94c5b88192dd43f09c692d0535553)

(3) Use the calculated bounding box to obtain the drawn_area instead of calculating it for each drawn pixel with set_and_check_rect (commit https://github.com/pygame-community/pygame-ce/commit/2e3585308911a5c92ac2d89a95c690214f3a9e10) -- NOTE fails some tests, was fixed in the commit after

(5) Bypass set_at to avoid checking surface clip rect and check it when calculating the bounding box instead (commit https://github.com/pygame-community/pygame-ce/commit/bef7f717c44391b37bff85533125a8d2234e5fe1) -- Fixes issues with previous commit

(6) commit a7b3cd3b82e5193d447cd46e7760cc2952c2e05e, redo algorithm somewhat to use faster circle line drawing techniques